refactor(sidebar): split file tree panel
