+++ src/skills/default-layout/sidebar/fileTreeActions.ts
@@ -1,0 +1,259 @@
+import {
+  useCallback,
+  useMemo,
+  useState,
+  type MouseEvent,
+  type RefObject,
+} from "react";
+import { invoke } from "@tauri-apps/api/core";
+
+import type { ContextMenuItem } from "../../../components/primitives/context-menu";
+import {
+  parentDirOf,
+  type ContextMenuState,
+  type TreeNode,
+} from "./fileTreeModel";
+
+interface UseFileTreeActionsArgs {
+  onEvent: (eventType: string, payload?: unknown) => void;
+  projectPath: string;
+  projectPathRef: RefObject<string>;
+  refreshFolder: (folderPath: string) => Promise<void>;
+}
+
+function useFileTreeActions({
+  onEvent,
+  projectPath,
+  projectPathRef,
+  refreshFolder,
+}: UseFileTreeActionsArgs) {
+  const [contextMenu, setContextMenu] = useState<ContextMenuState | null>(null);
+  const closeContextMenu = useCallback(() => setContextMenu(null), []);
+
+  const openContextMenu = useCallback((e: MouseEvent, node: TreeNode) => {
+    e.preventDefault();
+    e.stopPropagation();
+    // Raw client coords; the ContextMenu primitive clamps + corrects
+    // for WebKit zoom-frame drift before positioning.
+    setContextMenu({ x: e.clientX, y: e.clientY, node });
+  }, []);
+
+  const onContextNewFile = useCallback(async () => {
+    if (!contextMenu) return;
+    const parentPath = parentDirOf(contextMenu.node);
+    closeContextMenu();
+    const name = window.prompt("New file name");
+    if (!name) return;
+    const target = `${parentPath.replace(/\/$/, "")}/${name}`;
+    try {
+      await invoke("fs_create_file", {
+        root: projectPathRef.current,
+        path: target,
+      });
+      await refreshFolder(parentPath);
+      onEvent("file-tree-open", {
+        filePath: target,
+        rootPath: projectPathRef.current,
+      });
+    } catch (err) {
+      window.alert(`Failed to create file: ${String(err)}`);
+    }
+  }, [closeContextMenu, contextMenu, onEvent, projectPathRef, refreshFolder]);
+
+  const onContextNewFolder = useCallback(async () => {
+    if (!contextMenu) return;
+    const parentPath = parentDirOf(contextMenu.node);
+    closeContextMenu();
+    const name = window.prompt("New folder name");
+    if (!name) return;
+    const target = `${parentPath.replace(/\/$/, "")}/${name}`;
+    try {
+      await invoke("fs_create_dir", {
+        root: projectPathRef.current,
+        path: target,
+      });
+      await refreshFolder(parentPath);
+    } catch (err) {
+      window.alert(`Failed to create folder: ${String(err)}`);
+    }
+  }, [closeContextMenu, contextMenu, projectPathRef, refreshFolder]);
+
+  const onContextRename = useCallback(async () => {
+    if (!contextMenu) return;
+    const node = contextMenu.node;
+    closeContextMenu();
+    const name = window.prompt("Rename to", node.entry.name);
+    if (!name || name === node.entry.name) return;
+    const dirIdx = Math.max(
+      node.entry.path.lastIndexOf("/"),
+      node.entry.path.lastIndexOf("\\"),
+    );
+    const parentPath =
+      dirIdx >= 0 ? node.entry.path.slice(0, dirIdx) : node.entry.path;
+    const target = `${parentPath}/${name}`;
+    try {
+      await invoke("fs_rename", {
+        root: projectPathRef.current,
+        from: node.entry.path,
+        to: target,
+      });
+      onEvent("file-tree-rename", {
+        from: node.entry.path,
+        to: target,
+        kind: node.entry.kind,
+      });
+      await refreshFolder(parentPath);
+    } catch (err) {
+      window.alert(`Rename failed: ${String(err)}`);
+    }
+  }, [closeContextMenu, contextMenu, onEvent, projectPathRef, refreshFolder]);
+
+  const onContextDelete = useCallback(async () => {
+    if (!contextMenu) return;
+    const node = contextMenu.node;
+    closeContextMenu();
+    if (!window.confirm(`Move "${node.entry.name}" to the trash?`)) return;
+    const dirIdx = Math.max(
+      node.entry.path.lastIndexOf("/"),
+      node.entry.path.lastIndexOf("\\"),
+    );
+    const parentPath =
+      dirIdx >= 0 ? node.entry.path.slice(0, dirIdx) : node.entry.path;
+    try {
+      await invoke("fs_delete", {
+        root: projectPathRef.current,
+        path: node.entry.path,
+      });
+      onEvent("file-tree-delete", {
+        path: node.entry.path,
+        kind: node.entry.kind,
+      });
+      await refreshFolder(parentPath);
+    } catch (err) {
+      window.alert(`Delete failed: ${String(err)}`);
+    }
+  }, [closeContextMenu, contextMenu, onEvent, projectPathRef, refreshFolder]);
+
+  const onContextCopyPath = useCallback(async () => {
+    if (!contextMenu) return;
+    const path = contextMenu.node.entry.path;
+    closeContextMenu();
+    try {
+      await navigator.clipboard.writeText(path);
+    } catch {
+      window.alert(path);
+    }
+  }, [closeContextMenu, contextMenu]);
+
+  const onContextCopyRelativePath = useCallback(async () => {
+    if (!contextMenu) return;
+    const root = projectPathRef.current.replace(/\/+$/, "");
+    const path = contextMenu.node.entry.path;
+    const rel = path.startsWith(root + "/")
+      ? path.slice(root.length + 1)
+      : path;
+    closeContextMenu();
+    try {
+      await navigator.clipboard.writeText(rel);
+    } catch {
+      window.alert(rel);
+    }
+  }, [closeContextMenu, contextMenu, projectPathRef]);
+
+  const onContextRevealInFinder = useCallback(async () => {
+    if (!contextMenu) return;
+    const path = contextMenu.node.entry.path;
+    closeContextMenu();
+    try {
+      await invoke("fs_reveal_in_file_manager", {
+        root: projectPath,
+        path,
+      });
+    } catch (err) {
+      window.alert(`Reveal failed: ${String(err)}`);
+    }
+  }, [closeContextMenu, contextMenu, projectPath]);
+
+  const onContextOpenWithDefault = useCallback(async () => {
+    if (!contextMenu) return;
+    const path = contextMenu.node.entry.path;
+    closeContextMenu();
+    try {
+      await invoke("fs_open_in_default_app", {
+        root: projectPath,
+        path,
+      });
+    } catch (err) {
+      window.alert(`Open failed: ${String(err)}`);
+    }
+  }, [closeContextMenu, contextMenu, projectPath]);
+
+  const fileTreeMenuItems: ContextMenuItem[] = useMemo(
+    () =>
+      contextMenu
+        ? [
+            {
+              id: "new-file",
+              label: "New File…",
+              onSelect: onContextNewFile,
+            },
+            {
+              id: "new-folder",
+              label: "New Folder…",
+              onSelect: onContextNewFolder,
+            },
+            { type: "separator" },
+            {
+              id: "reveal-in-finder",
+              label: "Reveal in File Manager",
+              onSelect: onContextRevealInFinder,
+            },
+            {
+              id: "open-with-default",
+              label: "Open with default app",
+              disabled: contextMenu.node.entry.kind !== "file",
+              onSelect: onContextOpenWithDefault,
+            },
+            { type: "separator" },
+            { id: "rename", label: "Rename…", onSelect: onContextRename },
+            {
+              id: "delete",
+              label: "Move to Trash…",
+              danger: true,
+              onSelect: onContextDelete,
+            },
+            { type: "separator" },
+            {
+              id: "copy-path",
+              label: "Copy Path",
+              onSelect: onContextCopyPath,
+            },
+            {
+              id: "copy-rel",
+              label: "Copy Relative Path",
+              onSelect: onContextCopyRelativePath,
+            },
+          ]
+        : [],
+    [
+      contextMenu,
+      onContextCopyPath,
+      onContextCopyRelativePath,
+      onContextDelete,
+      onContextNewFile,
+      onContextNewFolder,
+      onContextOpenWithDefault,
+      onContextRename,
+      onContextRevealInFinder,
+    ],
+  );
+
+  return {
+    contextMenu,
+    fileTreeMenuItems,
+    openContextMenu,
+    setContextMenu,
+  };
+}
+
+export { useFileTreeActions };
